feat: Remove exception integration
